Replace std::auto_ptr with std::unique_ptr
This is a proof of concept to test with at this point. The c++11
deprecation of auto_ptr kicks out a lot of noise in the build process,
which of course, hides real warnings.

This patch should work on any compiler and stdlib that support the
basics of c++11's move semantics. Initial smoke testing doesn't seem to
cause instability, but only limited testing has been done so far.

When this project moves to the c++14 standard on lecacy platforms, there
is an opportunity to cut out a bit of code in app.cpp by using
std::make_unique<>() instead of the idiom used there that remains from
the poor design of std::auto_ptr.